refactor: moved AudioArray and AudioBus
