+++ packages/react-native-audio-api/android/src/main/cpp/audioapi/android/core/AudioDecoder.cpp
@@ -1,5 +1,5 @@
-#include <audioapi/core/utils/AudioArray.h>
-#include <audioapi/core/utils/AudioBus.h>
+#include <audioapi/utils/AudioArray.h>
+#include <audioapi/utils/AudioBus.h>
 #include <audioapi/core/utils/AudioDecoder.h>
 
 #include <android/log.h>
