+++ packages/react-native-audio-api/common/cpp/audioapi/core/destinations/AudioDestinationNode.cpp
@@ -1,7 +1,7 @@
 #include <audioapi/core/AudioNode.h>
 #include <audioapi/core/BaseAudioContext.h>
 #include <audioapi/core/destinations/AudioDestinationNode.h>
-#include <audioapi/core/utils/AudioBus.h>
+#include <audioapi/utils/AudioBus.h>
 #include <audioapi/core/utils/AudioNodeManager.h>
 
 namespace audioapi {
