+++ packages/react-native-audio-api/common/cpp/audioapi/core/sources/AudioBuffer.cpp
@@ -1,6 +1,6 @@
 #include <audioapi/core/sources/AudioBuffer.h>
-#include <audioapi/core/utils/AudioArray.h>
-#include <audioapi/core/utils/AudioBus.h>
+#include <audioapi/utils/AudioArray.h>
+#include <audioapi/utils/AudioBus.h>
 
 #include <utility>
